Moved the bluetooth logic to it's own class BLEClient which the Mower class builds on top of.
